+++ src/content/components/DictPanel/index.tsx
@@ -1,24 +1,35 @@
 import React from 'react'
 import { DictionariesState } from '../../redux/modules/dictionaries'
-import { AppConfig, DictID, DictConfigs } from '@/app-config'
-import { SelectionInfo } from '@/_helpers/selection'
+import { AppConfig, DictID, DictConfigs, MtaAutoUnfold } from '@/app-config'
+import { SelectionInfo, getDefaultSelectionInfo } from '@/_helpers/selection'
 import { MsgSelection } from '@/typings/message'
 import { Omit } from '@/typings/helpers'
+
+import CSSTransition from 'react-transition-group/CSSTransition'
+
+import { translate } from 'react-i18next'
+import { TranslationFunction } from 'i18next'
 
 import MenuBar, { MenuBarProps, MenuBarDispatchers } from '../MenuBar'
 import DictItem, { DictItemProps, DictItemDispatchers } from '../DictItem'
 
+const isSaladictPopupPage = !!window.__SALADICT_POPUP_PAGE__
+
 export type DictPanelDispatchers = DictItemDispatchers & MenuBarDispatchers & {
-  searchText: (arg?: { id?: DictID, info?: SelectionInfo | string }) => any
+  readonly searchText: (arg?: { id?: DictID, info?: SelectionInfo | string }) => any
+  readonly searchBoxUpdate: (arg: { text: string, index: number }) => any
+  readonly updateItemHeight: (id: DictID | '_mtabox', height: number) => any
 }
 
 type ChildrenProps =
   DictPanelDispatchers &
   Omit<MenuBarProps,
+    't' |
     'searchHistory' |
     'activeDicts'
   > &
   Omit<DictItemProps,
+    't' |
     'id' |
     'text' |
     'dictURL' |
@@ -29,15 +40,112 @@ type ChildrenProps =
 
 export interface DictPanelProps extends ChildrenProps {
   readonly isAnimation: boolean
+  readonly panelMaxHeightRatio: number
+  readonly mtaAutoUnfold: MtaAutoUnfold
   readonly dictionaries: DictionariesState['dictionaries']
   readonly allDictsConfig: DictConfigs
   readonly langCode: AppConfig['langCode']
   readonly selection: MsgSelection
 }
 
-export default class DictPanel extends React.Component<DictPanelProps> {
+interface DictPanelState {
+  mtaBoxHeight: number
+}
+
+export class DictPanel extends React.Component<DictPanelProps & { t: TranslationFunction }, DictPanelState> {
+  MtaBoxRef = React.createRef<HTMLTextAreaElement>()
+
+  state = {
+    mtaBoxHeight: 0
+  }
+
+  searchText = (arg?: { id?: DictID, info?: SelectionInfo | string }) => {
+    if (this.state.mtaBoxHeight !== 0) {
+      const { mtaAutoUnfold } = this.props
+      if (!mtaAutoUnfold ||
+          mtaAutoUnfold === 'once' ||
+          mtaAutoUnfold === 'popup' && !isSaladictPopupPage
+      ) {
+        this.setState({ mtaBoxHeight: 0 })
+      }
+    }
+    return this.props.searchText(arg)
+  }
+
+  toggleMtaBox = (e?: React.MouseEvent<HTMLButtonElement>) => {
+    if (e) { e.currentTarget.blur() }
+    this.setState(preState => {
+      return { mtaBoxHeight: preState.mtaBoxHeight <= 0
+        ? window.innerHeight * this.props.panelMaxHeightRatio * 0.4
+        : 0
+      }
+    })
+  }
+
+  handleMtaBoxInput = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+    this.props.searchBoxUpdate({
+      index: this.props.searchBoxIndex,
+      text: e.currentTarget.value
+    })
+  }
+
+  handleMtaBoxKeyDown = (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
+    if (e.key === 'Enter' && e.ctrlKey) {
+      e.preventDefault()
+      if (this.props.searchBoxText) {
+        return this.searchText({
+          info: getDefaultSelectionInfo({
+            text: this.props.searchBoxText,
+            title: this.props.t('fromSaladict'),
+            favicon: 'https://raw.githubusercontent.com/crimx/ext-saladict/dev/public/static/icon-16.png'
+          }),
+        })
+      }
+    }
+  }
+
+  componentDidMount () {
+    const { mtaAutoUnfold } = this.props
+    if (mtaAutoUnfold === 'once' ||
+        mtaAutoUnfold === 'always' ||
+        mtaAutoUnfold === 'popup' && isSaladictPopupPage
+    ) {
+      this.toggleMtaBox()
+    }
+  }
+
+  componentDidUpdate (prevProps: DictPanelProps, prevState: DictPanelState) {
+    if (prevState.mtaBoxHeight <= 0 &&
+        this.state.mtaBoxHeight > 0 &&
+        this.MtaBoxRef.current
+    ) {
+      this.MtaBoxRef.current.focus()
+    }
+
+    if (prevState.mtaBoxHeight !== this.state.mtaBoxHeight) {
+      this.props.updateItemHeight('_mtabox', this.state.mtaBoxHeight)
+    }
+
+    if (Boolean(prevProps.mtaAutoUnfold) !== Boolean(this.props.mtaAutoUnfold)) {
+      if (this.props.mtaAutoUnfold !== 'popup' || isSaladictPopupPage) {
+        this.toggleMtaBox()
+      }
+    }
+  }
+
+  renderMtaBox = () => (
+    <textarea
+      ref={this.MtaBoxRef}
+      value={this.props.searchBoxText}
+      onChange={this.handleMtaBoxInput}
+      onKeyDown={this.handleMtaBoxKeyDown}
+      style={{ fontSize: this.props.fontSize }}
+    />
+  )
+
   render () {
     const {
+      t,
       activeConfigID,
       configProfiles,
       isAnimation,
@@ -46,7 +154,9 @@ export default class DictPanel extends React.Component<DictPanelProps> {
       langCode,
       handleDragAreaMouseDown,
       handleDragAreaTouchStart,
-      searchText,
+      searchBoxUpdate,
+      searchBoxText,
+      searchBoxIndex,
       requestFavWord,
       shareImg,
       panelPinSwitch,
@@ -63,6 +173,10 @@ export default class DictPanel extends React.Component<DictPanelProps> {
     } = this.props
 
     const {
+      mtaBoxHeight,
+    } = this.state
+
+    const {
       dicts: dictsInfo,
       active: activeDicts,
     } = dictionaries
@@ -70,6 +184,7 @@ export default class DictPanel extends React.Component<DictPanelProps> {
     return (
       <div className={`panel-Root${isAnimation ? ' isAnimate' : ''}`}>
         {React.createElement(MenuBar, {
+          t,
           activeConfigID,
           configProfiles,
           isFav,
@@ -78,13 +193,34 @@ export default class DictPanel extends React.Component<DictPanelProps> {
           activeDicts: dictionaries.active,
           handleDragAreaMouseDown,
           handleDragAreaTouchStart,
-          searchText,
+          searchText: this.searchText,
+          searchBoxUpdate,
+          searchBoxText,
+          searchBoxIndex,
           requestFavWord,
           shareImg,
           panelPinSwitch,
           closePanel,
         })}
         <div className='panel-DictContainer'>
+          <div className='panel-MtaBox' style={{ height: mtaBoxHeight }}>
+            <CSSTransition
+              classNames='panel-MtaBoxTrans'
+              in={mtaBoxHeight > 0}
+              timeout={500}
+              mountOnEnter={true}
+              unmountOnExit={true}
+            >
+              {this.renderMtaBox}
+            </CSSTransition>
+          </div>
+          <button className='panel-MtaBoxBtn' onClick={this.toggleMtaBox}>
+            <svg width='10' height='10' viewBox='0 0 59.414 59.414' xmlns='http://www.w3.org/2000/svg'
+             className={'panel-MtaBoxBtn_Arrow' + (mtaBoxHeight > 0 ? ' isActive' : '')}
+            >
+              <path d='M58 14.146L29.707 42.44 1.414 14.145 0 15.56 29.707 45.27 59.414 15.56' />
+            </svg>
+          </button>
           {activeDicts.map(id => {
             let dictURL = allDictsConfig[id].page
             if (typeof dictURL !== 'string') {
@@ -92,6 +228,7 @@ export default class DictPanel extends React.Component<DictPanelProps> {
             }
 
             return React.createElement(DictItem, {
+              t,
               key: id,
               id,
               text: (dictionaries.searchHistory[0] || selection.selectionInfo).text,
@@ -101,7 +238,7 @@ export default class DictPanel extends React.Component<DictPanelProps> {
               panelWidth,
               searchStatus: (dictsInfo[id] as any).searchStatus,
               searchResult: (dictsInfo[id] as any).searchResult,
-              searchText,
+              searchText: this.searchText,
               updateItemHeight,
             })
           })}
@@ -110,3 +247,5 @@ export default class DictPanel extends React.Component<DictPanelProps> {
     )
   }
 }
+
+export default translate()(DictPanel)